[FEAT] cart: authen with api

+++ service/Cart.ts
@@ -2,9 +2,13 @@ import { makeRequest } from "./makeRequest";
 
 export function addProductToCart (uid : number | undefined , pid : number, amount: number) {
     if (typeof uid === 'number') {
+        const userToken = localStorage.getItem('token')
         return makeRequest(`/addcart/${uid}`, {
             method: "POST",
-            data: { pid, amount }
+            data: { pid, amount },
+            headers: {
+                'Authorization': `Bearer ${userToken}` 
+            }
         })
     }
 }
@@ -14,8 +18,15 @@ export function getProductCartByUser (uid : number | undefined) {
 }
 
 export function deleteProductCartById (uid: number | undefined, pid: number | undefined) {
-    return makeRequest(`/cart/${uid}`, {
-        method: "DELETE",
-        data: { pid }
-    })
+    if (typeof uid === 'number') {
+        const userToken = localStorage.getItem('token')
+        return makeRequest(`/cart/${uid}`, {
+            method: "DELETE",
+            data: { pid },
+            headers: {
+                'Authorization': `Bearer ${userToken}` 
+            }
+        })
+    }
+    
 }
